feat(RequestLogging): Include route and session
This returns the following optional `contextFields`:

- `route`
- `routeName`
- `x-session-id`

These are only set where present, so they shouldn't unnecessarily bloat
existing applications that don't make use of `@koa/router` or a session
header.

Koala has tried to remain middleware-agnostic to date, but these fields
are optional, and `@koa/router` is the clear router of choice in the Koa
ecosystem.