fix(ErrorMiddleware): Only respect certain `err.status`es (#199)
BREAKING CHANGE: Some HTTP clients throw errors with a `status` property to indicate the status code of the HTTP response. If such an error was not handled by a middleware or controller further up the chain, `ErrorMiddleware.handle` would previously pick up the error and reflect its status code back to your client.

Most of the time, this is not what you want. For example, if your server depends on an upstream service-to-service endpoint and it starts to respond with HTTP 401s, that may imply that your server is not appropriately authenticated and is at fault, and should default to a HTTP 500. This is now the default behaviour of `ErrorMiddleware.handle`.

If you want to throw an error that `ErrorMiddleware.handle` will pick up to modify your response status code, you are now limited to three options:

1. Use `http-errors`

   ```typescript
   import createError from 'http-errors';

   const controller = () => {
     throw new createError.ImATeapot();
   };
   ```

2. Use `ctx.throw`

   ```typescript
   const controller = (ctx) => {
     ctx.throw(418, 'Badness!');
   };
   ```

3. Include `isJsonResponse` and `status` properties on your error

   ```typescript
   class MyCustomError extends Error {
     constructor(
       message: string,
       public isJsonResponse: boolean,
       public status: number,
     ) {
       super(message);

       this.isJsonResponse = isJsonResponse;
       this.status = status;
     }
   }

   const controller = () => {
     throw new MyCustomError('Badness!', false, 418);
   };
   ```